% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_svclone.R
\name{analyze_svclone}
\alias{analyze_svclone}
\title{Process output from SVclone}
\usage{
analyze_svclone(samp, exper, final_truth, sv_info, tolerance = 100)
}
\arguments{
\item{samp}{an object of class 'character'. This object describes sample name.}

\item{exper}{an object of class 'character'. This object describes experiment number.}

\item{final_truth}{an object of class 'data frame'. This object stores the output from SVCFit.}

\item{sv_info}{an object of class 'data frame'. This object stores output from \code{parse_sv_info}}

\item{tolerance}{an object of class 'integer'. This object describes the limit of
base pair difference between svclone and svcfit output to be considered the same event}
}
\value{
sth
}
\description{
Process output from SVclone
}
