% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_bnd.R
\name{annotate_bnd}
\alias{annotate_bnd}
\title{Title}
\usage{
annotate_bnd(bnd_tmp, del)
}
\arguments{
\item{bnd_tmp}{an object of class 'data frame' which stores the sv input from \code{preproc_bnd}.}

\item{del}{an object of class 'data frame' which stores the sv input from \code{process_del}.}
}
\value{
sth
}
\description{
Title
}
