% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cnv.R
\name{assign_cnv}
\alias{assign_cnv}
\title{assign an SV to a CNV based on genomic region overlap}
\usage{
assign_cnv(sv_sum, cnv)
}
\arguments{
\item{sv_sum}{an object of class 'data frame'. This object stores the output from \code{sum_sv_info}.}

\item{cnv}{an object of class 'data frame'. This object stores the cnv data loaded from \code{load_data}.}
}
\value{
sth
}
\description{
assign an SV to a CNV based on genomic region overlap
}
