% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_svids.R
\name{assign_svids}
\alias{assign_svids}
\title{assign each SNP to an SV}
\usage{
assign_svids(sv_phase, sv_info, flank_snp = 500)
}
\arguments{
\item{sv_phase}{an object of class 'data frame'. This object stores the output from \code{parse_snp_on_sv}.}

\item{sv_info}{an object of class 'data frame'. This object stores the output from \code{parse_sv_info}.}

\item{flank_snp}{an object of class 'integer'. This object describes the upper limit of location difference
for an SNP to be assigned to an SV}
}
\value{
A data.frame with columns \code{snp_id} (row index into
\code{sv_phase}), \code{ID} (SV identifier from \code{sv_info}), and
\code{region} (\code{"start"} or \code{"end"}, indicating which SV
breakpoint the SNP was assigned to).
}
\description{
assign each SNP to an SV
}
