% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_clone.R
\name{attach_clone}
\alias{attach_clone}
\title{attach_clone is used to assign each structural variant to a clone designed in simulation. Only applicable if the ground truth is known and the simulation has multiple clones.}
\usage{
attach_clone(dat, truth, tolerance = 6)
}
\arguments{
\item{dat}{an object of class "dataframe". This object is a annotated vcf file storing structural variants to be compared}

\item{truth}{an object of class "dataframe". This object stores the clone assignment for each structural variants designed in a simulation}

\item{tolerance}{an object of class "integer". This variable sets the threshold of coordinates overlaps for two SVs}
}
\value{
add one more column "clone_num" to the dat
}
\description{
attach_clone is used to assign each structural variant to a clone designed in simulation. Only applicable if the ground truth is known and the simulation has multiple clones.
}
\examples{
example
}
