% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_svcf.R
\name{calculate_svcf}
\alias{calculate_svcf}
\title{Title}
\usage{
calculate_svcf(input, tumor_only = FALSE)
}
\arguments{
\item{input}{an object of class "dataframe". This object stores the set of information for structural variants used to calculate SVCF}

\item{tumor_only}{an object of class "Boolean". This variable indicates whether the vcf files were created under tumor-only mode}
}
\value{
an annotated vcf file
}
\description{
Title
}
