% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_overlap.R
\name{check_overlap}
\alias{check_overlap}
\title{check_overlap compare two sets of coordinates of structural variants and merge the structural variants with coordinates within tolerance.}
\usage{
check_overlap(dat, compare = NA, tolerance = 6, window = 1000)
}
\arguments{
\item{dat}{an object of class "dataframe". This object stores the set of
structural variants to be self-compared}

\item{tolerance}{an object of class "integer". This variable sets the threshold
coordinates overlaps for two SVs}

\item{window}{an object of class "Integer". This variable sets how many SVs
should be compared at the same time}
}
\value{
an filtered annotated vcf file
}
\description{
check_overlap compare two sets of coordinates of structural variants and merge the structural variants with coordinates within tolerance.
}
