% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_enumeration.R
\name{filterEdgesBasedOnCCFs}
\alias{filterEdgesBasedOnCCFs}
\title{Filter possible edges based on lineage precedence}
\usage{
filterEdgesBasedOnCCFs(graph_G, mcf, thresh = 0.1)
}
\description{
Filter possible edges based on lineage precedence
}
