% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_het_snp.R
\name{parse_het_snps}
\alias{parse_het_snps}
\title{parse information for germline heterozygous snps}
\usage{
parse_het_snps(het_snp)
}
\arguments{
\item{het_snp}{an object of class 'data frame'. This object stores the het snp data loaded from \code{load_data}.}
}
\value{
A data.frame with columns \code{CHROM}, \code{POS}, \code{REF},
\code{ALT}, \code{snp_ref} (reference-allele read count), \code{snp_alt}
(alternate-allele read count), and \code{dep} (total sequencing depth).
Restricted to single-nucleotide SNPs (length-1 REF and ALT alleles).
}
\description{
parse information for germline heterozygous snps
}
