% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{pop}
\alias{pop}
\title{Return the first row of an edge tibble}
\usage{
pop(edges_tb)
}
\arguments{
\item{edges_tb}{data.frame or tibble. The edge table to pop from.}
}
\description{
Return the first row of an edge tibble
}
\keyword{internal}
