% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_bnd.R
\name{preproc_bnd}
\alias{preproc_bnd}
\title{Process translocation}
\usage{
preproc_bnd(sv)
}
\arguments{
\item{sv}{an object of class 'data frame' which stores the sv input from \code{load_data}.}
}
\value{
sth
}
\description{
Process translocation
}
