% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_bnd.R
\name{process_del}
\alias{process_del}
\title{Title}
\usage{
process_del(sv, bnd_tmp, flank_del = 50)
}
\arguments{
\item{sv}{an object of class 'data frame' which stores the sv input from \code{load_data}.}

\item{bnd_tmp}{an object of class 'data frame' which is the output from \code{preproc_bnd}.}

\item{flank_del}{an object of class 'integer' which describes the maximum genomic
location difference between a deletion and translocation to be considered related.}
}
\value{
sth
}
\description{
Title
}
