% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clone.R
\name{read_clone}
\alias{read_clone}
\title{This function reads the ground truth designed for a simulation}
\usage{
read_clone(fp)
}
\arguments{
\item{fp}{an onject of class "Character". This variable is a path to bed files storing structural variants information}
}
\value{
a dataframe
}
\description{
This function reads the ground truth designed for a simulation
}
\examples{
example
}
