% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_sv_info.R
\name{sum_sv_info}
\alias{sum_sv_info}
\title{sumarise SV information (zygosity, phasing, reads)}
\usage{
sum_sv_info(sv_phase, assign_id, sv_info)
}
\arguments{
\item{sv_phase}{an object of class 'data frame'. This object stores the output from \code{parse_snp_on_sv}.}

\item{assign_id}{an object of class 'data frame'. This object stores the output from \code{assign_svids}.}

\item{sv_info}{an object of class 'data frame'. This object stores the output from \code{parse_sv_info}.}
}
\value{
A data.frame joining phasing and zygosity summaries with full SV
metadata.  Contains all \code{sv_info} columns plus \code{snp_alt},
\code{snp_ref}, \code{dep}, \code{onsv_alt}, \code{onsv_ref},
\code{zygosity_snp}, \code{sv_phase}, \code{no_snp} (1 if no SNPs were
available), and \code{zygosity}.
}
\description{
sumarise SV information (zygosity, phasing, reads)
}
