% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVCF.R
\name{SVCF}
\alias{SVCF}
\title{Calculate structural variant cellular fraction from vcf files}
\usage{
SVCF(
  vcf_path,
  overlap = TRUE,
  tolerance = 6,
  window = 100,
  multiple = FALSE,
  tumor_only = FALSE,
  truth_path = NULL,
  mode = "heritage",
  length_threshold = 0
)
}
\arguments{
\item{vcf_path}{an object of class "Character". This variable is the path to vcf files}

\item{overlap}{an object of class "Boolean". This variable indicates if a structural variants should be filtered based on coordinates overlap}

\item{tolerance}{an object of class "Integer". This variable sets the threshold coordinates overlaps for two SVs}

\item{window}{an object of class "Integer". This variable sets how many SVs should be compared at the same time}

\item{multiple}{an object of class "Boolean". This variable indicates whether the sample has multiple clone (only used for simulated data)}

\item{tumor_only}{an object of class "Boolean". This variable indicates whether the vcf files were created under tumor-only mode}

\item{truth_path}{an onject of class "Character". This variable is a path to bed files storing true structural variants information with clonal assignment. Each bed file should be named like "c1.bed, c2.bed" etc. Structural variants should be saved in a seperate bed file if they belong to different (sub)clone.}

\item{mode}{an onject of class "Character". This variable describe how true clonal information is saved. In "heritage" mode, bed files for all children clone contains all ancestral structural variants of their parents. In "separate" mode, children clones don't contain any ancestral structural variants.}

\item{length_threshold}{an object of class "integer". This variable set lowest threshold on the size of a structural variants}
}
\value{
an annotated vcf file
}
\description{
Calculate structural variant cellular fraction from vcf files
}
\examples{
example
}
