% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVCFit-package.R
\docType{package}
\name{SVCFit-package}
\alias{SVCFit}
\alias{SVCFit-package}
\title{SVCFit: Calculate Structural Variant Cellular Fraction}
\description{
Estimates the Structural Variant Cellular Fraction (SVCF) for inversions, deletions, tandem duplications, and translocations by integrating structural variant calls with allele-specific copy number variation and heterozygous SNP data. Includes a Dirichlet-process Gaussian mixture model clustering step and phylogenetic tree inference for longitudinal paired samples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KarchinLab/SVCFit}
  \item Report bugs at \url{https://github.com/KarchinLab/SVCFit/issues}
}

}
\author{
\strong{Maintainer}: Yunzhou Liu \email{yliu498@jh.edu}

}
