% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVCFit.R
\name{SVCFit}
\alias{SVCFit}
\title{Calculate structural variant cellular fraction from vcf files}
\usage{
SVCFit(
  vcf_path,
  simulation = FALSE,
  overlap = TRUE,
  tolerance = 6,
  window = 100,
  multiple = FALSE,
  tumor_only = FALSE,
  truth_path = NULL,
  length_filter = 0
)
}
\arguments{
\item{vcf_path}{an object of class "Character". This variable is the path to vcf files}

\item{simulation}{an object of class "Boolean". This variable indicates if the vcf files are from a simulation}

\item{overlap}{an object of class "Boolean". This variable indicates if a structural variants should be filtered based on coordinates overlap}

\item{tolerance}{an object of class "Integer". This variable sets the threshold coordinates overlaps for two SVs}

\item{window}{an object of class "Integer". This variable sets how many SVs should be compared at the same time}

\item{multiple}{an object of class "Boolean". This variable indicates whether the sample has multiple clone (only used for simulated data)}

\item{tumor_only}{an object of class "Boolean". This variable indicates whether the vcf files were created under tumor-only mode}

\item{truth_path}{an object of class "Character". This variable is the path to the ground truth designed for simulation}

\item{length_filter}{an object of class "integer". This variable set lowest threshold on the size of a structural variants}
}
\value{
an annotated vcf file
}
\description{
Calculate structural variant cellular fraction from vcf files
}
\examples{
example
}
