% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcfit_w_svc.R
\name{svcfit_w_svc}
\alias{svcfit_w_svc}
\title{Run SVCFit with reads from SVclone}
\usage{
svcfit_w_svc(svc, final_truth, thresh, samp, exper)
}
\arguments{
\item{svc}{an object of class 'data frame'. This object stores the output from \code{analyze_svclone}.}

\item{final_truth}{an object of class 'data frame'. This object stores the output from \code{calc_svcf}.}

\item{thresh}{an object of class 'numeric'. This object is strictly between 0 and 1, and describes the
threshold for SVCFit to decide whether an SV occurs before CNV or after.}

\item{exper}{an object of class 'character'. This object describes the experiment number.}

\item{sampan}{object of class 'character'. This object describes the sample name.}
}
\value{
sth
}
\description{
Run SVCFit with reads from SVclone
}
