% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{main wrapper to run all functions}
\usage{
analyze(samp, exper, thresh = 0.1, chr_lst = NULL)
}
\arguments{
\item{exper}{an object of class 'character'. This object describes the experiment number.}

\item{thresh}{an object of class 'numeric'. This object is strictly between 0 and 1, and describes the
threshold for SVCFit to decide whether an SV occurs before CNV or after.}

\item{sampan}{object of class 'character'. This object describes the sample name.}
}
\value{
sth
}
\description{
main wrapper to run all functions
}
