% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_truth.R
\name{attach_truth}
\alias{attach_truth}
\title{attach truth of simulation to the output of SVCFit}
\usage{
attach_truth(final, truth)
}
\arguments{
\item{final}{an object of class 'data frame'. This object stores the output from \code{analyze}.}

\item{truth}{an object of class 'data frame'. This object stores the output from \code{load_truth}.}
}
\value{
sth
}
\description{
attach truth of simulation to the output of SVCFit
}
