% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_svcf.R
\name{calc_svcf}
\alias{calc_svcf}
\title{Calculate SVCF}
\usage{
calc_svcf(anno_sv_cnv, sv_info, thresh = 0.1, samp, exper)
}
\arguments{
\item{anno_sv_cnv}{an object of class 'data frame'. This object stores the output from \code{annotate_cnv}.}

\item{sv_info}{an object of class 'data frame'. This object stores the output from \code{parse_sv_info}.}

\item{thresh}{an object of class 'numeric'. This object is strictly between 0 and 1, and describes the
threshold for SVCFit to decide whether an SV occurs before CNV or after.}

\item{samp}{an object of class 'character'. This object describes the sample name.}

\item{exper}{an object of class 'character'. This object describes the experiment number.}
}
\value{
sth
}
\description{
Calculate SVCF
}
