% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterize_sv.R
\name{characterize_sv}
\alias{characterize_sv}
\title{Characterize SV}
\usage{
characterize_sv(sv_phase, sv_info, cnv, flank_snp = 500, flank_cnv = 1000)
}
\arguments{
\item{sv_phase}{object of class 'dataframe'. This object stores the SV phasing
and zygosity information obtained from \emph{parse_snp_on_sv()}.}

\item{sv_info}{object of class 'dataframe'. This object stores SV genomic locations
and types obtained from \emph{parse_sv_info()}.}

\item{cnv}{object of class 'dataframe'. This object stores CNV information
obtained from \emph{load_data()}.}

\item{flank_snp}{Numeric. Flanking window (bp) used when mapping SNPs to SVs
in \code{\link{assign_svids}}. Default \code{500}.}

\item{flank_cnv}{Numeric. Flanking window (bp) used when assigning background
CNV in \code{\link{assign_background_cnv}}. Default \code{1000}.}
}
\value{
A data.frame combining CNV annotation from \code{\link{annotate_cnv}},
full SV metadata, and background CNV classification from
\code{\link{assign_background_cnv}}.  Key added columns: \code{cn_type},
\code{ASCN}, \code{bkg_cnv} (\code{"DUP"}, \code{"norm"}, or \code{"DEL"}
indicating the background copy number state flanking each SV breakpoint).
}
\description{
Characterize SV
}
