% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterize_sv.R
\name{characterize_sv}
\alias{characterize_sv}
\title{Characterize SV}
\usage{
characterize_sv(sv_phase, sv_info, cnv, flank_snp = 500, flank_cnv = 1000)
}
\arguments{
\item{sv_phase}{object of class 'dataframe'. This object stores the SV phasing
and zygosity information obtained from \emph{parse_snp_on_sv()}.}

\item{sv_info}{object of class 'dataframe'. This object stores SV genomic locations
and types obtained from \emph{parse_sv_info()}.}

\item{cnv}{object of class 'dataframe'. This object stores CNV information
obtained from \emph{load_data()}.}
}
\value{
sth
}
\description{
Characterize SV
}
