% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_data}
\alias{cluster_data}
\title{Perform full clustering pipeline on paired data}
\usage{
cluster_data(
  pair_path,
  pur_path,
  Kmax = 10,
  n_steps = 100,
  thr_min_w = 0.01,
  random_state = 0L,
  concentration = 1,
  min_n = 5,
  pair_num = 1
)
}
\description{
This preserves the original clustering path and only changes the
final mapping back to original SV rows.
}
