% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{edgesToAmLong}
\alias{edgesToAmLong}
\title{Convert edge list to long adjacency matrix with annotations}
\usage{
edgesToAmLong(edges)
}
\description{
Convert edge list to long adjacency matrix with annotations
}
