% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info.R
\name{extract_info}
\alias{extract_info}
\title{extract information}
\usage{
extract_info(
  p_het,
  p_onsv,
  p_sv,
  p_cnv,
  flank_del = 50,
  QUAL_tresh = 100,
  min_alt = 2,
  tumor_only = FALSE
)
}
\arguments{
\item{p_het}{object of class 'character'. This object stores the file path
to vcf file for heterozygous SNPs.}

\item{p_onsv}{object of class 'character'. This object stores the file path
to vcf files for heterozygous SNPs that are on SV supporting reads.}

\item{p_sv}{object of class 'character'. This object stores the file path
to vcf files for SVs.}

\item{p_cnv}{object of class 'character'. This object stores the file path
to CNV file.}

\item{flank_del}{object of class 'numeric'. This object describes the maximum allowed
differences in genomic locations for a deletion to be considered as overlapping to a translocation.}

\item{QUAL_tresh}{object of class 'numeric'. This object describes the minimum quality score
allowed to include an SV.}

\item{min_alt}{object of class 'numeric'. This object describes the minimum
amount of SV supporting read counts to include an SV.}

\item{tumor_only}{object of class 'Boolean'. This object describes if the SVs were called only using tumor BAM}
}
\value{
sth
}
\description{
extract information
}
