% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{load input data}
\usage{
load_data(p_het, p_onsv, p_sv, p_cnv, chr = NULL, tumor_only = FALSE)
}
\arguments{
\item{p_het}{object of class 'character'. This object stores the file path
to vcf file for heterozygous SNPs.}

\item{p_onsv}{object of class 'character'. This object stores the file path
to vcf files for heterozygous SNPs that are on SV supporting reads.}

\item{p_sv}{object of class 'character'. This object stores the file path
to vcf files for SVs.}

\item{p_cnv}{object of class 'character'. This object stores the file path
to CNV file.}

\item{chr}{an object of class 'character'. This object describes what chromosome to include for analysis.}

\item{tumor_only}{object of class 'Boolean'. This object describes if the SVs were called only using tumor BAM}
}
\value{
sth
}
\description{
load input data
}
