% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{load input data}
\usage{
load_data(
  samp,
  exper,
  base_dir = "~/Documents/Karchin_lab/visor/m_benchmark",
  chr = NULL
)
}
\arguments{
\item{samp}{an object of class 'character'. This object describes the sample name.}

\item{exper}{an object of class 'character'. This object describes the experiment number.}

\item{base_dir}{an object of class 'character'. This object describes the path where file is stored.}

\item{chr}{an object of class 'character'. This object describes what chromosome to include for analysis.}
}
\value{
sth
}
\description{
load input data
}
