% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_truth.R
\name{load_truth}
\alias{load_truth}
\title{load simulation truth data}
\usage{
load_truth(exper)
}
\arguments{
\item{exper}{an object of class 'character'. This object describes the experiment number.}
}
\value{
sth
}
\description{
load simulation truth data
}
