% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_het_snp.R
\name{parse_het_snps}
\alias{parse_het_snps}
\title{parse information for germline heterozygous snps}
\usage{
parse_het_snps(het_snp)
}
\arguments{
\item{het_snp}{an object of class 'data frame'. This object stores the het snp data loaded from \code{load_data}.}
}
\value{
sth
}
\description{
parse information for germline heterozygous snps
}
