% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_snp_on_sv.R
\name{parse_snp_on_sv}
\alias{parse_snp_on_sv}
\title{parse information for germline het snp that are close to SVs}
\usage{
parse_snp_on_sv(het_on_sv, snp_df)
}
\arguments{
\item{het_on_sv}{an object of class 'data frame'. This object stores the het snp on sv data loaded from \code{load_data}.}

\item{snp_df}{an object of class 'data frame'. This object stores the output from \code{parse_het_snps}.}
}
\value{
sth
}
\description{
parse information for germline het snp that are close to SVs
}
