% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{pre_process_cluster}
\alias{pre_process_cluster}
\title{Load and process data for clustering}
\usage{
pre_process_cluster(pair_path, pur_path)
}
\arguments{
\item{pair_path}{Character. Path to a file containing paired sample IDs (columns: pre_BAT, on_BAT).}

\item{pur_path}{Character. Path to a file containing purity estimates for each sample.}
}
\value{
A data frame containing processed CCF values for pre- and on-treatment samples
(\code{pre_BAT}, \code{on_BAT}), ready for input into clustering functions.
}
\description{
Reads paired sample IDs and purity files, calculates Cancer Cell Fraction (CCF)
for Structural Variants (SVs), and reshapes the data (wide format) for longitudinal clustering.
It handles shared and non-shared variants between pre- and on-treatment samples.
}
