% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_enumeration.R
\name{prepareGraph}
\alias{prepareGraph}
\title{Create tibble of possible edges from CCF values based on w_mat only}
\usage{
prepareGraph(mcf_mat, thresh)
}
\description{
Create tibble of possible edges from CCF values based on w_mat only
}
