% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_bnd.R
\name{proc_bnd}
\alias{proc_bnd}
\title{Process translocation}
\usage{
proc_bnd(sv, flank_del = 50)
}
\arguments{
\item{sv}{an object of class 'data frame' which stores the sv input from \code{load_data}.}

\item{flank_del}{an object of class 'integer' which describes the maximum genomic location
difference between a deletion and a translocation to be considered overlapping}
}
\value{
sth
}
\description{
Process translocation
}
